修改公式

+++ S_1/Assets/HalfWorld/Scripts/GameUnit/Field/FieldData.cs
@@ -7,14 +7,14 @@ namespace ELGame
 	[System.Serializable]
     public struct FieldData
     {
-		public float resVolume;		//总共资源量
-		public float resRemain; 	//剩余资源量
-		public int difficulty;		//难度值
-		public float expRate;				//给予的经验
-		public float goldRate;			//给予的金钱
-		public float fameRate;			//给予的声望
+		public float resVolume;		    //总共资源量
+		public float resRemain; 	    //剩余资源量
+		public int difficulty;		    //难度值
+		public int expRate;			//给予的经验
+		public int goldRate;			//给予的金钱
+		public int fameRate;			//给予的声望
 
-		public FieldData(float resVolume, float expRate, float goldRate, float fameRate, int difficulty)
+		public FieldData(float resVolume, int expRate, int goldRate, int fameRate, int difficulty)
 		{
 			this.resVolume = resVolume;
 			this.resRemain = this.resVolume;
